Drop vboot-utils build dependency
The package contains tools to build and sign FIT (Flat Image Tree) images
that are used in u-boot based Chromebooks. But it isn't really needded to
build u-boot, since FIT images are not shipped in uboot-images-armv{7,8}.

Also, the vboot-utils package has as build dependency the trousers-devel
package and this gets pulled as a dependency to build u-boot as well.

But the trousers package is really deprecated since TPM 1.2 chips aren't
used anymore nowadays. Let's just drop that unneeded build dependency.